requirements.txt
gradio
pandas
plotly.express
specklepy